+++ codegen/src/markdown/table_of_contents.rs
@@ -4,6 +4,9 @@ pub fn generate(path: &std::path::Path) -> Result<String, CodegenError> {
     let output = std::process::Command::new("npx")
         .arg("--yes")
         .arg("markdown-toc")
+        .arg("--no-firsth1")
+        .arg("--bullets")
+        .arg("-")
         .arg(path)
         .output()?;