+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. Dates are d
 
 #### [Unreleased](https://github.com/hougesen/mdsf/compare/v0.9.0...HEAD)
 
+- feat(tools): add support for cue [`#962`](https://github.com/hougesen/mdsf/pull/962)
 - feat(tools): add support for atlas [`#961`](https://github.com/hougesen/mdsf/pull/961)
 - chore: update version to dev [`020a67a`](https://github.com/hougesen/mdsf/commit/020a67a65be83dba53c6ac69d68cea3bd9b953a8)